Initial Kubeflow support
If the environment variable `NB_PREFIX` is used, it is assumed rstudio is from kubeflow and thus userconf.sh will set `www-root-path` to
 `$NB_PREFIX`, set `www-port=8888` and it will disable auth. `RSTUDIO_VERSION=daily` or `RSTUDIO_VERSION=preview` need to be used (daily not working, preview not tested). On daily, the 404 page loads saying:
```
RSudio Server
The requested page was not found.
Please contact your system administrator for assistance, or click here to go back.
```
This was tested with a Dockerfile that exposes port 8888.