Set WORKDIR for *all* images
Without this, trying to run these for anything other than RStudio
puts us in `/` rather than in `/home/rstudio`.

```bash
$ docker run -it rocker/geospatial:4.2.2 pwd
/
`

RStudio puts the user in /home/rstudio as expected (probably something
to do with PAM, I assume). This change makes the behavior of using the
images without RStudio match that of using it with.