binder: Install packages required for Jupyter to PDF convert
Per https://github.com/jupyter/nbconvert/issues/1328, these
are the packages needed for Jupyter to be able to convert
to PDF. Without it, you get unfriendly errors like [this]
(https://github.com/2i2c-org/infrastructure/issues/3188#issuecomment-1755969247).

Given that the binder image includes jupyter, and texlive is included
in the base, I hope it would be reasonable for the binder image
to include enough texlive packages for PDF conversion to work.
Otherwise, it works in RStudio but not in Jupyter.