Install python into a venv (#718)
- Continues to use python and venv from Ubuntu LTS repositories, so they
are supported as with everything else that is gotten from apt (see
https://github.com/rocker-org/rocker-versioned2/issues/670#issuecomment-1619748511)
- Doesn't currently change any permissions, so present behavior is
preserved. However, in the future, we should probably change ownership
so end users can install packages in there at runtime (see
https://github.com/rocker-org/rocker-versioned2/issues/670#issuecomment-1624691981)
- Sets the VIRTUAL_ENV environment variable to path of the venv we
create. This is what the `source activate` script does, and Reticulate
also looks for this to discover which python to use (see point 4 of
https://rstudio.github.io/reticulate/articles/versions.html#order-of-discovery)
- Sets up PATH appropriately, so python and python3 refer to what is in
our venv. This, along with the previous step, ensures same behavior as
users typing `source ${VIRTUAL_ENV}/bin/activate` without actually
having to do that, preserving end user behavioral semantics. See
https://github.com/rocker-org/rocker-versioned2/issues/670#issuecomment-1601860904
- RStudio is also told about new `PATH` and `VIRTUAL_ENV`, using the
same pattern as `install_texlive.sh`
- Remove the explicit symlink of python3 -> python, as venv handles this
automatically.
- `install_python.sh` now needs to be `source`d, following same pattern
as `install_texlive.sh`

Decisions to be made:

- Where do we set the appropriate env variables (VIRTUAL_ENV and PATH)?
They need to be set for `install_python.sh` to work correctly. I've set
them in the binder image for now, but it should probably be set on a
more base image. This is a no-op if `install-python.sh` is not called
anywhere.

TODO:
- [x] Update `NEWS` (can be done once everything else is finalized)

Ref https://github.com/rocker-org/rocker-versioned2/issues/670

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>